Opprydding
Blitt kvitt en del warnings i kompileringen
Refaktorert AAD-funksjoner ut av Kontrollservice
