+++ src/main/kotlin/no/nav/meldeplikt/meldekortservice/api/Weblogic.kt
@@ -10,15 +10,18 @@ import no.nav.meldeplikt.meldekortservice.utils.WEBLOGIC_PING_PATH
 import no.nav.meldeplikt.meldekortservice.utils.respondOrError
 import no.nav.meldeplikt.meldekortservice.utils.swagger.*
 
+@io.ktor.locations.KtorExperimentalLocationsAPI
 fun Routing.weblogicApi() {
     pingWeblogic()
 }
 
 @Group("Weblogic")
 @Location("$WEBLOGIC_PING_PATH")
+@io.ktor.locations.KtorExperimentalLocationsAPI
 class PingWeblogicInput
 
 // Endepunkt for ping av weblogic
+@io.ktor.locations.KtorExperimentalLocationsAPI
 fun Routing.pingWeblogic() =
     get<PingWeblogicInput>(
         "Ping av weblogic for å sjekke at Arena er oppe".responds(
